Add files via upload
This project aims to create a web application to visualize graph of
Wireless Sensors Network (WSN) configuration stored in a database. The
application should be able to read from the database the information
regarding sensors’ locations, connections, and other parameters, and present
it in a graphical form.
The user should be able to interact with the graph, adjust parameters,
and view the results. The application will use an existing database to store
the WSN configuration, JS/HTML/CSS to create a graphical representation
of the WSN configuration, and a user friendly interface allowing the user to
view the graph and interact with it.
User can give the csv file and the file then store in database . from
database the graph is generated .The data from the database is display and
user can download the image of the graph .